src/requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
Jinja2==2.11.2
MarkupSafe==1.1.1
PyYAML==5.3.1
requests==2.23.0
urllib3==1.25.9
